Move PhysicalPlanner to `physical_planer` module
